+++ src/models/Settings.php
@@ -2,31 +2,39 @@
 
 namespace mutation\translate\models;
 
+use Craft;
 use craft\base\Model;
 
 class Settings extends Model
 {
     public $pluginName = 'Translations';
+    public $sourceLanguage = '';
     public $categories = [['category' => 'site']];
     public $addMissingTranslations = true;
     public $addMissingSiteRequestOnly = true;
     public $excludedMessages = [];
 
+    public function defineRules(): array
+    {
+        return [
+            [['sourceLanguage', 'categories'], 'required'],
+        ];
+    }
+
+    public function getSourceLanguage()
+    {
+        return $this->sourceLanguage ?: Craft::$app->getSites()->getPrimarySite()->language;
+    }
+
     public function getCategories()
     {
-        $cats = [];
-        foreach ($this->categories as $cat) {
-            if (isset($cat['category'])) {
-                $cats[] = $cat['category'];
-            } else {
-                $cats[] = $cat;
-            }
-        }
-        return $cats;
+        if (!$this->categories) return ['site'];
+        return collect($this->categories)->flatten()->toArray();
     }
 
     public function getExcludedMessages()
     {
-        return $this->excludedMessages;
+        if (!$this->excludedMessages) return [];
+        return collect($this->excludedMessages)->flatten()->toArray();
     }
 }